detector/requirements.txt
uvicorn>=0.29.0
opencv-python==4.9.0.80
numpy==1.26.4
ffmpeg-python>=0.2.0
requests>=2.31.0
fastapi>=0.110.0
lapx
python-multipart>=0.0.9
kafka-python==2.0.2
